refactor: move virtual list into another component
