+++ packages/app/src/features/NoteEditor/NoteVersions/NoteVersions.tsx
@@ -1,27 +1,16 @@
-import React, { useCallback, useEffect, useRef, useState } from 'react';
-import { Trans, useTranslation } from 'react-i18next';
+import React, { useCallback, useEffect, useState } from 'react';
+import { useTranslation } from 'react-i18next';
 import { FaEraser, FaFloppyDisk } from 'react-icons/fa6';
 import { LOCALE_NAMESPACE } from 'src/i18n';
 import { WorkspaceEvents } from '@api/events/workspace';
-import { Box, Button, HStack, Stack, Switch, Text, VStack } from '@chakra-ui/react';
+import { Box, Button, HStack, Switch, Text, VStack } from '@chakra-ui/react';
 import { BoxWithCenteredContent } from '@components/BoxWithCenteredContent';
 import { TextWithIcon } from '@components/TextWithIcon';
 import { NoteVersion } from '@core/features/notes/history/NoteVersions';
-import { TELEMETRY_EVENT_NAME } from '@core/features/telemetry';
 import { useEventBus, useNotesHistory } from '@features/App/Workspace/WorkspaceProvider';
-import { useMemoizedCallback } from '@features/MainScreen/TagsPanel/useMemoizedCallback';
-import { useTelemetryTracker } from '@features/telemetry';
 import { useConfirmDialog } from '@hooks/useConfirmDialog';
-import { useVirtualizer } from '@tanstack/react-virtual';
 
-import { NoteVersionItem } from './NoteVersionItem';
-
-const ONE_DAY_IN_MS = 1000 * 60 * 60 * 24;
-const timestampToDays = (ms: number) => Math.floor(ms / ONE_DAY_IN_MS);
-const getTimestampAgeInDays = (ms: number) => timestampToDays(Date.now() - ms);
-
-export const formatNoteVersionPreview = (version: NoteVersion) =>
-	`${new Date(version.createdAt).toLocaleString()} (${version.text.length} chars)`;
+import { NoteVersionsList } from './NoteVersionsList';
 
 // TODO: implement lazy loading
 export const NoteVersions = ({
@@ -45,7 +34,6 @@ export const NoteVersions = ({
 	isReadOnly?: boolean;
 }) => {
 	const { t } = useTranslation(LOCALE_NAMESPACE.features);
-	const telemetry = useTelemetryTracker();
 	const noteHistory = useNotesHistory();
 
 	const [versions, setVersions] = useState<NoteVersion[] | null>(null);
@@ -69,154 +57,6 @@ export const NoteVersions = ({
 	}, [eventBus, updateVersionsList]);
 
 	const confirm = useConfirmDialog();
-	const listRootRef = useRef<HTMLDivElement>(null);
-
-	const onApply = useMemoizedCallback(
-		useCallback(
-			(version: NoteVersion) => {
-				const applyVersion = () => {
-					onVersionApply(version);
-					telemetry.track(TELEMETRY_EVENT_NAME.NOTE_VERSION_APPLIED, {
-						versionAgeInDays: getTimestampAgeInDays(version.createdAt),
-					});
-				};
-
-				confirm(({ onClose }) => ({
-					title: t('note.versions.confirmApply.title'),
-					content: (
-						<VStack gap="1rem" align="start">
-							<Text>
-								<Trans
-									i18nKey="note.versions.confirmApply.description"
-									ns={LOCALE_NAMESPACE.features}
-									values={{
-										version: formatNoteVersionPreview(version),
-									}}
-									components={{
-										secondary: <Text as="span" variant="secondary" />,
-									}}
-								/>
-							</Text>
-							<Text>{t('note.versions.confirmApply.warning')}</Text>
-						</VStack>
-					),
-					action: (
-						<>
-							<Button
-								variant="accent"
-								onClick={() => {
-									applyVersion();
-									onClose();
-								}}
-							>
-								{t('note.versions.confirmApply.apply')}
-							</Button>
-							<Button
-								variant="accent"
-								onClick={() => {
-									onShowVersion(version);
-									onClose();
-								}}
-							>
-								{t('note.versions.confirmApply.preview')}
-							</Button>
-							<Button onClick={onClose}>
-								{t('common:actions.cancel')}
-							</Button>
-						</>
-					),
-				}));
-			},
-			[confirm, onShowVersion, onVersionApply, t, telemetry],
-		),
-	);
-
-	const onShow = useMemoizedCallback(
-		useCallback(
-			(version: NoteVersion) => {
-				console.log('onShow click');
-				onShowVersion(version);
-
-				telemetry.track(TELEMETRY_EVENT_NAME.NOTE_VERSION_VIEWED, {
-					versionAgeInDays: getTimestampAgeInDays(version.createdAt),
-				});
-			},
-			[onShowVersion, telemetry],
-		),
-	);
-
-	const onDelete = useMemoizedCallback(
-		useCallback(
-			(version: NoteVersion) => {
-				const deleteVersion = async () => {
-					await noteHistory.delete([version.id]);
-					eventBus.emit(WorkspaceEvents.NOTE_HISTORY_UPDATED, noteId);
-
-					telemetry.track(TELEMETRY_EVENT_NAME.NOTE_VERSION_DELETED, {
-						versionAgeInDays: getTimestampAgeInDays(version.createdAt),
-					});
-				};
-
-				confirm(({ onClose }) => ({
-					title: t('note.versions.confirmDelete.title'),
-					content: (
-						<VStack gap="1rem" align="start">
-							<Text>
-								<Trans
-									i18nKey="note.versions.confirmDelete.description"
-									ns={LOCALE_NAMESPACE.features}
-									values={{
-										version: formatNoteVersionPreview(version),
-									}}
-									components={{
-										secondary: <Text as="span" variant="secondary" />,
-									}}
-								/>
-							</Text>
-							<Text>{t('note.versions.confirmDelete.warning')}</Text>
-						</VStack>
-					),
-					action: (
-						<>
-							<Button
-								variant="accent"
-								onClick={() => {
-									deleteVersion();
-									onClose();
-								}}
-							>
-								{t('note.versions.confirmDelete.delete')}
-							</Button>
-							<Button
-								variant="accent"
-								onClick={() => {
-									onShowVersion(version);
-									onClose();
-								}}
-							>
-								{t('note.versions.confirmDelete.preview')}
-							</Button>
-							<Button onClick={onClose}>
-								{t('common:actions.cancel')}
-							</Button>
-						</>
-					),
-				}));
-			},
-			[confirm, eventBus, noteHistory, noteId, onShowVersion, t, telemetry],
-		),
-	);
-
-	// FIXME: https://github.com/TanStack/virtual/issues/1119
-	// eslint-disable-next-line react-hooks/incompatible-library
-	const virtualizer = useVirtualizer({
-		count: versions?.length ?? 0,
-		getScrollElement: () => listRootRef.current,
-		estimateSize: () => 20,
-		overscan: 6,
-		useFlushSync: false,
-		useCachedMeasurements: true,
-	});
 
 	return (
 		<VStack w="100%" maxH="100%">
@@ -299,14 +139,7 @@ export const NoteVersions = ({
 					</label>
 				</HStack>
 			</HStack>
-			<Box
-				ref={listRootRef}
-				width="100%"
-				overflow="auto"
-				display="flex"
-				flex={1}
-				flexFlow="column"
-			>
+			<Box width="100%" overflow="auto" display="flex" flex={1} flexFlow="column">
 				{versions && versions.length === 0 && (
 					<BoxWithCenteredContent>
 						<Text fontSize="1.3rem">{t('note.versions.empty')}</Text>
@@ -318,32 +151,13 @@ export const NoteVersions = ({
 					</BoxWithCenteredContent>
 				)}
 				{versions !== null && versions.length > 0 && (
-					<Stack
-						width="100%"
-						flexShrink={0}
-						gap={0}
-						style={{
-							height: virtualizer.getTotalSize(),
-							paddingTop: virtualizer.getVirtualItems()[0]?.start ?? 0,
-						}}
-					>
-						{virtualizer.getVirtualItems().map((virtualRow) => {
-							const version = versions[virtualRow.index];
-
-							return (
-								<NoteVersionItem
-									ref={virtualizer.measureElement}
-									key={virtualRow.index}
-									data-index={virtualRow.index}
-									version={version}
-									onApply={onApply(version)}
-									onPreview={onShow(version)}
-									onDelete={onDelete(version)}
-									isReadOnly={isReadOnly}
-								/>
-							);
-						})}
-					</Stack>
+					<NoteVersionsList
+						noteId={noteId}
+						versions={versions}
+						onShowVersion={onShowVersion}
+						onVersionApply={onVersionApply}
+						isReadOnly={isReadOnly}
+					/>
 				)}
 			</Box>
 		</VStack>
